+++ sql/internal/drivertest/doc.go
@@ -1,0 +1,2 @@
+// Package drivertest contains a common test suite for all SQL drivers.
+package drivertest