+++ sql/mysql/doc.go
@@ -1,0 +1,3 @@
+// Package mysql contains an SQL projection driver for MySQL and compatible
+// databases such as MariaDB.
+package mysql