+++ sql/postgres/doc.go
@@ -1,0 +1,2 @@
+// Package postgres contains an SQL projection driver for PostgreSQL.
+package postgres