+++ sql/sqlite/doc.go
@@ -1,0 +1,2 @@
+// Package sqlite contains an SQL projection driver for SQLite v3.
+package sqlite